+++ src/test/java/pl/coffeepower/DynatraceGenericLogIngestAppenderIntegrationTest.java
@@ -9,16 +9,21 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import java.io.IOException;
+import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
+import java.time.Duration;
 import java.util.List;
+import java.util.UUID;
 
 import org.apache.http.HttpStatus;
+import org.awaitility.Awaitility;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
+import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.core.WireMockConfiguration;
@@ -29,6 +34,8 @@ import com.google.common.collect.Lists;
 
 class DynatraceGenericLogIngestAppenderIntegrationTest {
 
+	private static final URL CONFIG_RESOURCE =
+			DynatraceGenericLogIngestAppenderIntegrationTest.class.getClassLoader().getResource("log4j2.xml");
 	private static WireMockServer SERVER;
 
 	private final ObjectMapper objectMapper = new ObjectMapper();
@@ -50,8 +57,7 @@ class DynatraceGenericLogIngestAppenderIntegrationTest {
 		Path tempConfigFile = Files.createTempFile("log4j", "test");
 		assertThat(
 				Files.copy(
-						DynatraceGenericLogIngestAppenderIntegrationTest.class.getClassLoader().getResource("log4j2.xml")
-								.openStream(),
+						CONFIG_RESOURCE.openStream(),
 						tempConfigFile,
 						StandardCopyOption.REPLACE_EXISTING))
 				.isGreaterThan(0L);
@@ -78,15 +84,25 @@ class DynatraceGenericLogIngestAppenderIntegrationTest {
 		} catch (Throwable t) {
 			fail(t);
 		} finally {
+			Awaitility.await()
+					.pollDelay(Duration.ofSeconds(1L))
+					.timeout(Duration.ofSeconds(5L))
+					.until(() -> !SERVER.getServeEvents().getRequests().isEmpty());
+
 			List<ServeEvent> requests = SERVER.getServeEvents().getRequests();
 
 			assertThat(requests)
 					.hasSize(5)
 					.extracting(event -> event.getRequest().getBodyAsString())
 					.extracting(objectMapper::readTree)
+					.allSatisfy(jsonNode -> assertThat(jsonNode.get("dt.trace_id"))
+							.isNotNull()
+							.extracting(JsonNode::asText)
+							.extracting(UUID::fromString)
+							.isNotNull())
 					.extracting(jsonNode -> Lists.newArrayList(jsonNode.fieldNames()))
 					.allSatisfy(fields -> assertThat(fields)
-							.containsOnlyOnce("timestamp", "level", "service.name", "dt.os.type", "message"));
+							.containsOnly("timestamp", "level", "service.name", "dt.os.type", "message", "dt.trace_id"));
 		}
 	}
 }